# app/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic API
  description: API for managing mechanics and their services.
  version: 1.0.0
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and return a customer specific token."
      parameters:
          - in: "body"
            name: "body"
            description: "Login credentials"
            required: true
            schema:
              $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "kjsbd87y23jsdhf8237y23jsdhf"
              message: "Login successful"
              status: "success"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Returns tickets"
      description: "Endpoint to retrieve all the customers tickets."
      security:
          - bearerAuth: []
      parameters:
          - in: path
            name: my-tickets
            description: "Information required to retrieve tickets"
            required: true
            schema:
              $ref: "#/definitions/TicketPayload"
      responses:
        200:
          description: "Tickets Retrieved "
          schema:
            $ref: "#/definitions/TicketResponse"
          examples:
            application/json:
              token: "kjsbd87y23jsdhf8237y23jsdhf"
              message: "Ticket retrievel successful"
              status: "success"
  /customers:
    post:
      tags:
        - Customers
      summary: "Endpoint to create/register a new customer"
      description: "Endpoint to create a new customer in our Library system, emails must be unique."
      parameters:
          - in: "body"
            name: "body"
            description: "Information required to create a new Customer"
            required: true
            schema:
              $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Succesfull Customer Creation"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: johnd@email.com
              password: password123
              phone: 123-456-7890

    get:
      tags:
        - Customers
      summary: "Returns a list of Customers"
      description: "Endpoint to retrieve a list of all customers."
      responses:
        200:
          description: "Retrieved all Customers Successfully"
          schema:
            $ref: "#/definitions/AllCustomersResponse"

    put:
      tags:
        - Customers
      summary: "Endpoint to update a customer."
      description: "Endpoint to update a customer, this is a token authentication route, and the customer must be logged in to update their account."
      security:
          - bearerAuth: []
      parameters:
          - in: "body"
            name: "body"
            description: "Information required to update a Customer"
            required: true
            schema:
              $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
            description: "Succesfull Customer Update"
            schema:
              $ref: "#/definitions/UpdateCustomerResponse"
            examples:
              application/json:
                id: 1
                name: Dohn Joe
                phone: "123-456-7896"
                email: dohnj@email.com
                password: password123

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete a customer."
      description: "Endpoint to delete a customer. Requires token authentication, customer must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Sucessfully Deleted customer"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer deleted."
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Returns a Customer"
      description: "Endpoint to retrieve a customer."
      parameters:
          - name: customer_id
            in: path
            required: true
            type: integer
            description: ID of the customer to retrieve
      responses:
        200:
          description: "Retrieved Customer Successfully"
          schema:
            $ref: "#/definitions/OneCustomerResponse"
          examples:
            application/json:
              id: 1
              name: Dohn Joe
              phone: 123-456-7892
              email: dohnj@email.com
              password: password123
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  TicketPayload:
    type: "object"
    properties:
      token:
        type: "string"

    required:
      - token
  TicketResponse:
    type: "object"
    properties:
      ticket id:
        type: "integer"
      mechanic id:
        type: "integer"
      customer id:
        type: "integer"
      service date:
        type: "date"
      service description:
      vin:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"

    required:
      - name
      - email
      - password
      - phone
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
  OneCustomerResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        phone:
          type: "string"
  AllCustomersResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        phone:
          type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"

    required:
      - name
      - email
      - password
      - phone
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
...